requirements.txt
opencv-python
numPy
pillow
tqdm
grpcio
protobuf
lpips
clean-fid